Adding in About Section
This is not perfect, as I want the ‘div’ background colour to wrap
around the whole time… Or the text to get bigger to keep the box full
as it gets bigger and smaller. Ultimately it would be good if the whole
background could be one block colour and this show up over the home
pages like one of the turn.js style sheets… Might be a work in progress
tho. Added in home navigation to the bar. Figure out how you have the
nav bar saved once and just refer to it again and again.