About adjusted
Hey - so I’ve got the text aligned so it can’t go off the page, but
this is not an ideal solution if we change things. But it’s responsive.

I’m having the same problem where I can’t actually make it stay in the
middle of the page. It might not need to for this section but it would
look better if it had a max width of x pixels (whatever the max width
of pixels should be). so in wide screen it doesn’t sit at the top of
the page.

Opposite problem of what I’m having with the Pepi logo on the main
page. But possibly because the Pepi logo is x px down from the top,
which as the page gets bigger the % is moving it off the page.

I think that when you added you put img into the css, with center,
center, fixed, which works well, but I don’t know how you do that for
text as well.

Wonder if this makes sense