All components there
Everything is up - just need to tweak a few things to make it look
prettier

The header is slightly left justified, need it to stretch across;

Add in a ‘home’ navigation function? Or just add in the navigation dots?

Where will we add in FAQ’s etc, should this all be on the motivated
signup page? as a microsite?

Can I add in the images that pop up like anyways one?